+++ application/ui/src/features/models/model-listing/model-variants/model-variant-table.component.tsx
@@ -24,8 +24,6 @@ export const ModelVariantTable = ({ model, format }: ModelVariantTableProps) => 
     return (
         <TableView aria-label={`Model variants for ${model.id}`} overflowMode={'wrap'} density={'compact'}>
             <TableHeader>
-                <Column isRowHeader>ARCHITECTURE</Column>
-                <Column isRowHeader>LICENSE</Column>
                 <Column isRowHeader>PRECISION</Column>
                 <Column isRowHeader>SIZE</Column>
                 <Column align='end'>
@@ -35,8 +33,6 @@ export const ModelVariantTable = ({ model, format }: ModelVariantTableProps) => 
             <TableBody items={variants}>
                 {(variant) => (
                     <Row key={`${variant.format}-${variant.precision}`}>
-                        <Cell>{model.architecture}</Cell>
-                        <Cell>Apache 2.0</Cell>
                         <Cell>{variant.precision.toUpperCase()}</Cell>
                         <Cell>{formatModelSize(variant.weights_size)}</Cell>
                         <Cell>
